introduce common TransactionBuilders
