+++ tck/src/main/java/com/hedera/hashgraph/tck/methods/sdk/AccountService.java
@@ -1,7 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
 package com.hedera.hashgraph.tck.methods.sdk;
 
-import com.google.protobuf.InvalidProtocolBufferException;
 import com.hedera.hashgraph.sdk.*;
 import com.hedera.hashgraph.tck.annotation.JSONRPC2Method;
 import com.hedera.hashgraph.tck.annotation.JSONRPC2Service;
@@ -10,10 +9,8 @@ import com.hedera.hashgraph.tck.methods.sdk.param.account.*;
 import com.hedera.hashgraph.tck.methods.sdk.param.transfer.*;
 import com.hedera.hashgraph.tck.methods.sdk.response.AccountAllowanceResponse;
 import com.hedera.hashgraph.tck.methods.sdk.response.AccountResponse;
-import com.hedera.hashgraph.tck.util.KeyUtils;
-import java.time.Duration;
+import com.hedera.hashgraph.tck.util.TransactionBuilders;
 import java.util.Map;
-import java.util.Optional;
 
 /**
  * AccountService for account related methods
@@ -28,41 +25,7 @@ public class AccountService extends AbstractJSONRPC2Service {
 
     @JSONRPC2Method("createAccount")
     public AccountResponse createAccount(final AccountCreateParams params) throws Exception {
-        AccountCreateTransaction accountCreateTransaction = new AccountCreateTransaction();
-        params.getKey().ifPresent(key -> {
-            try {
-                accountCreateTransaction.setKeyWithoutAlias(KeyUtils.getKeyFromString(key));
-            } catch (InvalidProtocolBufferException e) {
-                throw new IllegalArgumentException(e);
-            }
-        });
-
-        params.getInitialBalance()
-                .ifPresent(initialBalanceTinybars -> accountCreateTransaction.setInitialBalance(
-                        Hbar.from(Long.parseLong(initialBalanceTinybars), HbarUnit.TINYBAR)));
-
-        params.getReceiverSignatureRequired().ifPresent(accountCreateTransaction::setReceiverSignatureRequired);
-
-        params.getAutoRenewPeriod()
-                .ifPresent(autoRenewPeriodSeconds -> accountCreateTransaction.setAutoRenewPeriod(
-                        Duration.ofSeconds(Long.parseLong(autoRenewPeriodSeconds))));
-
-        params.getMemo().ifPresent(accountCreateTransaction::setAccountMemo);
-
-        params.getMaxAutoTokenAssociations()
-                .ifPresent(autoAssociations ->
-                        accountCreateTransaction.setMaxAutomaticTokenAssociations(autoAssociations.intValue()));
-
-        params.getStakedAccountId()
-                .ifPresent(stakedAccountId ->
-                        accountCreateTransaction.setStakedAccountId(AccountId.fromString(stakedAccountId)));
-
-        params.getStakedNodeId()
-                .ifPresent(stakedNodeId -> accountCreateTransaction.setStakedNodeId(Long.parseLong(stakedNodeId)));
-
-        params.getDeclineStakingReward().ifPresent(accountCreateTransaction::setDeclineStakingReward);
-
-        params.getAlias().ifPresent(accountCreateTransaction::setAlias);
+        AccountCreateTransaction accountCreateTransaction = TransactionBuilders.AccountBuilder.buildCreate(params);
 
         params.getCommonTransactionParams()
                 .ifPresent(commonTransactionParams ->
@@ -81,43 +44,7 @@ public class AccountService extends AbstractJSONRPC2Service {
 
     @JSONRPC2Method("updateAccount")
     public AccountResponse updateAccount(final AccountUpdateParams params) throws Exception {
-        AccountUpdateTransaction accountUpdateTransaction = new AccountUpdateTransaction();
-
-        params.getAccountId()
-                .ifPresent(accountId -> accountUpdateTransaction.setAccountId(AccountId.fromString(accountId)));
-
-        params.getKey().ifPresent(key -> {
-            try {
-                accountUpdateTransaction.setKey(KeyUtils.getKeyFromString(key));
-            } catch (InvalidProtocolBufferException e) {
-                throw new IllegalArgumentException(e);
-            }
-        });
-
-        params.getReceiverSignatureRequired().ifPresent(accountUpdateTransaction::setReceiverSignatureRequired);
-
-        params.getAutoRenewPeriod()
-                .ifPresent(autoRenewPeriodSeconds -> accountUpdateTransaction.setAutoRenewPeriod(
-                        Duration.ofSeconds(Long.parseLong(autoRenewPeriodSeconds))));
-
-        params.getMemo().ifPresent(accountUpdateTransaction::setAccountMemo);
-
-        params.getExpirationTime()
-                .ifPresent(expirationTime ->
-                        accountUpdateTransaction.setExpirationTime(Duration.ofSeconds(Long.parseLong(expirationTime))));
-
-        params.getMaxAutoTokenAssociations()
-                .ifPresent(autoAssociations ->
-                        accountUpdateTransaction.setMaxAutomaticTokenAssociations(autoAssociations.intValue()));
-
-        params.getStakedAccountId()
-                .ifPresent(stakedAccountId ->
-                        accountUpdateTransaction.setStakedAccountId(AccountId.fromString(stakedAccountId)));
-
-        params.getStakedNodeId()
-                .ifPresent(stakedNodeId -> accountUpdateTransaction.setStakedNodeId(Long.parseLong(stakedNodeId)));
-
-        params.getDeclineStakingReward().ifPresent(accountUpdateTransaction::setDeclineStakingReward);
+        AccountUpdateTransaction accountUpdateTransaction = TransactionBuilders.AccountBuilder.buildUpdate(params);
 
         params.getCommonTransactionParams()
                 .ifPresent(commonTransactionParams ->
@@ -131,13 +58,7 @@ public class AccountService extends AbstractJSONRPC2Service {
 
     @JSONRPC2Method("deleteAccount")
     public AccountResponse deleteAccount(final AccountDeleteParams params) throws Exception {
-        AccountDeleteTransaction accountDeleteTransaction = new AccountDeleteTransaction();
-
-        params.getDeleteAccountId()
-                .ifPresent(accountId -> accountDeleteTransaction.setAccountId(AccountId.fromString(accountId)));
-
-        params.getTransferAccountId()
-                .ifPresent(accountId -> accountDeleteTransaction.setTransferAccountId(AccountId.fromString(accountId)));
+        AccountDeleteTransaction accountDeleteTransaction = TransactionBuilders.AccountBuilder.buildDelete(params);
 
         params.getCommonTransactionParams()
                 .ifPresent(commonTransactionParams ->
@@ -151,9 +72,7 @@ public class AccountService extends AbstractJSONRPC2Service {
 
     @JSONRPC2Method("approveAllowance")
     public AccountAllowanceResponse approveAllowance(final AccountAllowanceParams params) throws Exception {
-        AccountAllowanceApproveTransaction tx = new AccountAllowanceApproveTransaction();
-
-        params.getAllowances().ifPresent(allowances -> allowances.forEach(allowance -> approve(tx, allowance)));
+        AccountAllowanceApproveTransaction tx = TransactionBuilders.AccountBuilder.buildApproveAllowance(params);
 
         params.getCommonTransactionParams()
                 .ifPresent(commonParams -> commonParams.fillOutTransaction(tx, sdkService.getClient()));
@@ -165,9 +84,7 @@ public class AccountService extends AbstractJSONRPC2Service {
 
     @JSONRPC2Method("deleteAllowance")
     public AccountAllowanceResponse deleteAllowance(final AccountAllowanceParams params) throws Exception {
-        AccountAllowanceDeleteTransaction tx = new AccountAllowanceDeleteTransaction();
-
-        params.getAllowances().ifPresent(allowances -> allowances.forEach(allowance -> delete(tx, allowance)));
+        AccountAllowanceDeleteTransaction tx = TransactionBuilders.AccountBuilder.buildDeleteAllowance(params);
 
         params.getCommonTransactionParams()
                 .ifPresent(commonParams -> commonParams.fillOutTransaction(tx, sdkService.getClient()));
@@ -186,10 +103,7 @@ public class AccountService extends AbstractJSONRPC2Service {
      */
     @JSONRPC2Method("transferCrypto")
     public Map<String, String> transferCrypto(final TransferCryptoParams params) throws Exception {
-        TransferTransaction transferTransaction = new TransferTransaction();
-
-        params.getTransfers()
-                .ifPresent(transfers -> transfers.forEach(txParams -> processTransfer(transferTransaction, txParams)));
+        TransferTransaction transferTransaction = TransactionBuilders.TransferBuilder.buildTransfer(params);
 
         params.getCommonTransactionParams()
                 .ifPresent(
@@ -294,54 +208,4 @@ public class AccountService extends AbstractJSONRPC2Service {
         });
     }
 
-    private void approve(AccountAllowanceApproveTransaction tx, AllowanceParams allowance) {
-        AccountId owner = AccountId.fromString(allowance.getOwnerAccountId().orElseThrow());
-        AccountId spender = AccountId.fromString(allowance.getSpenderAccountId().orElseThrow());
-
-        allowance
-                .getHbar()
-                .ifPresent(hbar ->
-                        tx.approveHbarAllowance(owner, spender, Hbar.fromTinybars(Long.parseLong(hbar.getAmount()))));
-
-        allowance
-                .getToken()
-                .ifPresent(token -> tx.approveTokenAllowance(
-                        TokenId.fromString(token.getTokenId()), owner, spender, token.getAmount()));
-
-        allowance.getNft().ifPresent(nft -> approveNFT(tx, owner, spender, nft));
-    }
-
-    private void delete(AccountAllowanceDeleteTransaction tx, AllowanceParams allowance) {
-        var owner = AccountId.fromString(allowance.getOwnerAccountId().orElseThrow());
-        var tokenId = allowance.getTokenId().orElseThrow();
-
-        if (allowance.getSerialNumbers().isPresent()) {
-            allowance.getSerialNumbers().get().forEach(serialNumber -> {
-                var nftId = new NftId(TokenId.fromString(tokenId), Long.parseLong(serialNumber));
-                tx.deleteAllTokenNftAllowances(nftId, owner);
-            });
-        }
-    }
-
-    private void approveNFT(
-            AccountAllowanceApproveTransaction tx,
-            AccountId owner,
-            AccountId spender,
-            AllowanceParams.TokenNftAllowance nft) {
-        TokenId tokenId = TokenId.fromString(nft.getTokenId());
-        Optional<String> delegateSpender = Optional.ofNullable(nft.getDelegatingSpender());
-
-        if (!nft.getSerialNumbers().isEmpty()) {
-            nft.getSerialNumbers().forEach(serial -> {
-                NftId nftId = new NftId(tokenId, serial);
-                delegateSpender.ifPresentOrElse(
-                        ds -> tx.approveTokenNftAllowance(nftId, owner, spender, AccountId.fromString(ds)),
-                        () -> tx.approveTokenNftAllowance(nftId, owner, spender));
-            });
-        } else if (nft.getAllSerials()) {
-            tx.approveTokenNftAllowanceAllSerials(tokenId, owner, spender);
-        } else {
-            tx.deleteTokenNftAllowanceAllSerials(tokenId, owner, spender);
-        }
-    }
 }
